Refactor to move resolver functions from utils to resolver package,
