+++ resolver/starwars/film.go
@@ -20,7 +20,7 @@ func NewFilmResolver(client *swapi.Client) *filmResolver {
 }
 
 func (*filmResolver) ID(ctx context.Context, f *swapi.Film) (string, error) {
-	return utils.ID(ctx, f.URL)
+	return ID(ctx, f.URL)
 }
 
 func (*filmResolver) Name(ctx context.Context, f *swapi.Film) (string, error) {
