+++ resolver/starwars/resource.go
@@ -1,28 +1,19 @@
-package utils
+package starwars
 
 import (
 	"context"
-	"fmt"
 	"strconv"
 
 	"gqlgen-starwars/errors"
 	"gqlgen-starwars/swapi"
+	"gqlgen-starwars/utils"
 )
-
-func ParseId(id string) (int, error) {
-	resourceId, err := strconv.Atoi(id)
-	if err != nil {
-		return 0, errors.NewUserInputError(fmt.Sprintf("Invalid id: %v", id), "id")
-	}
-
-	return resourceId, nil
-}
 
 // ID returns graphql ID field value for a url
 func ID(ctx context.Context, url string) (string, error) {
 	id, err := swapi.ResourceId(url)
 	if err != nil {
-		GetLogEntry(ctx).Error("Failed to extract id from resource url")
+		utils.GetLogEntry(ctx).Error("Failed to extract id from resource url")
 
 		return "", errors.NewParsingError(err)
 	}
